Refactor TensorBoard logging in Geneformer tests
### Description
This commit refactors the TensorBoard logging verification in the Geneformer test suite. It moves the TensorBoard availability check and log verification function to a shared module, updates the test parameters to ensure TensorBoard logging is consistently enabled, and modifies the expected metrics to align with recent changes.

### Type of changes
- [x] Refactor

### Pre-submit Checklist
- [x] I have tested these changes locally
- [x] I have added/updated tests as needed
- [x] All existing tests pass successfully
