+++ sub-packages/bionemo-testing/src/bionemo/testing/__init__.py
@@ -12,3 +12,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from bionemo.testing.tensorboard import TENSORBOARD_AVAILABLE, verify_tensorboard_logs
+
+__all__ = ["TENSORBOARD_AVAILABLE", "verify_tensorboard_logs"]
